Moved native_libraries.get_emba_db.py script to infoextractors.native_lib_file.py module
